requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
statsmodels>=0.14
python-dateutil>=2.8
mlflow>=2.12
psycopg2-binary>=2.9
